requirements.txt
streamlit
langchain 
langchain_groq 
langchain_core 
python-dotenv 
huggingface-hub 
sentence-transformers 
pypdf 
langchain_community 